Split model into Entities and Collections
