+++ src/Model/Collection/Collection.php
@@ -1,18 +1,33 @@
 <?php
-namespace TC\Model;
+
+namespace TC\Model\Collection;
+
+use TC\Lib\Database;
+
 
 /**
  * This file is part of the Simple PHP Framework
  *
- * Default Model class
+ * Default collection
  *
- * @category Models
+ * @category Model
  * @package  TC
  * @author   Manuel Tancoigne <m.tancoigne@gmail.com>
  * @license  http://www.opensource.org/licenses/mit-license.php MIT License
  * @link     https://github.com/tatooine-coders/simple-php-framework/
  */
-class Model
+class Collection
 {
 
+    public function fetchAll()
+    {
+    }
+
+    public function saveAll()
+    {
+    }
+
+    public function deleteAll()
+    {
+    }
 }
